Refactor feature tests
There are a number of shared steps between the banner and preferences
page feature tests so we've moved these into a renamed
`cucumber_cookie_management_test_steps` file to avoid duplication and
ambiguous definitions.  We've also added a feature test to cover the
behaviour where a user rejects all cookies in the banner and then
accepts them all on the preferences page.

Finally, we've refactored the scenario names for the preferences pages
feature tests

Co-authored-by: Stacy Hayes <stacy.hayes@citizensadvice.org.uk>