Update the cookies engine to use /cookies-preferences
Up until now the url for the cookies preference page has been
"/cookies-preferences/cookie_preference".  However, it was agreed that
we will flatten this to "/cookies-preferences" to avoid a repeat of
"/about-us/contact-us/contact-us/contact-us".

This commit updates the routes and all the various links to ensure they
point to the correct path