Move FormOption inside the engine's namespace
Because `isolate_namespace` is enabled in the engine's configuration,
any classes outside of the namespace aren't made available to the host
app.

This was fine for projects that already had their own version of
the `FormOption` model class, but when there wasn't one, it would throw
an error.

Moving `FormOption` under the `CitizensAdviceCookiePreferences`
namespace fixed the issue.