Add no-js cookie banner
We need to render a no-js version of the cookie banner for users that
don't have it enabled for any reason.  This involves adding the
`noscript` version of the banner into the erb file, and adding a new
`aria-hidden` attribute so that it won't be read out by screenreaders.
The cookie banner JS file has been updated to remove this attribute if a
user views the banner with JS enabled