Add unit tests to CookieManagement
Using an ActiveSupport::HashWithIndifferentAccess in place of the cookie
jar, which was proving difficult to use within the test context, we've
added some basic tests that give us confidence that further refactoring
hasn't broken anything.

Additionally, we've included `cookie_preference` and
`cookie_preference_set` in the list of essential cookies.

Co-authored-by: Leeky <robert.lee-cann@citizensadvice.org.uk>