Use meta-tags gem and update desc.
We have gotten the content through from Andrew for the description,
so have updated the code to use this and to use translations ready
for Welsh.

We have also updated demo to use the meta tags gem that the other
apps use to have consistent implementation of the tagging.

Co-authored by: Lucy Stringer <lucy.stringer@citizensadvice.org.uk>