Refactor link generation in banner and on preference page
We're able to access the request params in the helper method, which
means that we don't need to pass in the `country` as a param.  As such
we can return everything back to the yaml file, and just pass in the
result of the `how_we_use_cookies_url` method to ensure the correct url
is rendered

Additionally, we can also move the code to generate the preferences page
url into the helpers as again we have access to the country params.
This will avoid needing to make changes in the host apps, and keeps all
the changes in the engine