Test country specific behaviour
On the preference page we need to render different layout for EW to
Scotland, and different links for each country. This tests that behaviour.
We also need to conditionally render the country specific links to the
cookie info page.

On any page with the banner we need to conditionally render country specific
links for both the cookie info page and the cookie preference page.

Co-authored-by: Kath Pobee-Norris <kath.pobee-norris@citizensadvice.org.uk>