Use Addressable gem to normalize ReturnUrl param
We've been seeing a number of requests from AmazonBot, where the URI
string has been incorrectly encoded as UTF-8 (the HTTP standards require
URIs to always be ASCII encoded).

I initially considered catching the exception and ignoring the ReturnUrl
altogether, but then thought it was probably better to normalize the URI
(courtesy of the Addressable gem), so we try to do the right thing
despite the weird encoding.